chore: fix client
